requirements.txt
# FastAPI 및 웹 서버
fastapi==0.115.12
uvicorn==0.27.1
httpx
pydantic
python-multipart
jinja2
pytest

# 웹 스크래핑
playwright==1.52.0

# 스케줄링
apscheduler==3.10.4
pytz==2024.1

# 환경 변수 관리
python-dotenv==1.0.1

# 비디오 생성
moviepy==1.0.3

# YouTube API
google-api-python-client==2.118.0
oauth2client==4.1.3
httplib2==0.22.0
google-auth-oauthlib==1.2.0

# 이미지 처리
Pillow==10.2.0 